requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via bioservices
appnope==0.1.0            # via ipykernel, ipython
attrs==19.3.0             # via jsonschema
backcall==0.1.0           # via ipython
beautifulsoup4==4.8.2     # via bioservices
bioservices==1.6.0        # via -r requirements.in
bleach==3.1.0             # via nbconvert
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
colorama==0.4.3           # via easydev
colorlog==4.0.2           # via bioservices, easydev
cycler==0.10.0            # via matplotlib
decorator==4.4.1          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
easydev==0.9.38           # via bioservices
entrypoints==0.3          # via nbconvert
gevent==1.4.0             # via grequests
greenlet==0.4.15          # via gevent
grequests==0.4.0          # via bioservices
idna==2.8                 # via requests
importlib-metadata==1.3.0  # via jsonschema
ipykernel==5.1.3          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.11.0           # via -r requirements.in, ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
jedi==0.15.2              # via ipython
jinja2==2.10.3            # via nbconvert, notebook
jsonschema==3.2.0         # via nbformat
jupyter-client==5.3.4     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.0.0    # via jupyter
jupyter-core==4.6.1       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r requirements.in
kiwisolver==1.1.0         # via matplotlib
lxml==4.4.2               # via bioservices
markupsafe==1.1.1         # via jinja2
matplotlib-venn==0.11.5   # via -r requirements.in
matplotlib==3.1.2         # via -r requirements.in, matplotlib-venn
mistune==0.8.4            # via nbconvert
more-itertools==8.0.2     # via zipp
nbconvert==5.6.1          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==6.0.2           # via -r requirements.in, jupyter, widgetsnbextension
numpy==1.18.0             # via -r requirements.in, matplotlib, matplotlib-venn, pandas, scipy
pandas==0.25.3            # via -r requirements.in, bioservices
pandocfilters==1.4.2      # via nbconvert
parso==0.5.2              # via jedi
pexpect==4.7.0            # via easydev, ipython
pickleshare==0.7.5        # via ipython
prometheus-client==0.7.1  # via notebook
prompt-toolkit==2.0.10    # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
pygments==2.5.2           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.6          # via matplotlib
pyrsistent==0.15.6        # via jsonschema
python-dateutil==2.8.1    # via jupyter-client, matplotlib, pandas
python-igraph==0.7.1.post6  # via -r requirements.in
pytz==2019.3              # via pandas
pyzmq==18.1.1             # via jupyter-client, notebook
qtconsole==4.6.0          # via jupyter
requests-cache==0.5.2     # via bioservices
requests==2.22.0          # via bioservices, grequests, requests-cache
scipy==1.4.1              # via -r requirements.in, matplotlib-venn
send2trash==1.5.0         # via notebook
six==1.13.0               # via bleach, cycler, jsonschema, prompt-toolkit, pyrsistent, python-dateutil, traitlets
soupsieve==1.9.5          # via beautifulsoup4
suds-jurko==0.6           # via bioservices
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tornado==6.0.3            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.45.0              # via -r requirements.in
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
urllib3==1.25.7           # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
wrapt==1.11.2             # via bioservices
xlrd==1.2.0               # via -r requirements.in
xmltodict==0.12.0         # via bioservices
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
